requirements.txt
aiogram~=2.23.1
python-dotenv~=0.21.0
requests~=2.28.1
beautifulsoup4~=4.11.1
aiosqlite~=0.17.0
tortoise-orm